Fixed Sargon iOS example app. (Xcode 16) (#224)
* Fixed Sargon iOS example app.

* fix tests

* Revert "fix tests"

This reverts commit 79ce9798de254fafdf9c0155cec3d514deb92484.

+++ examples/iOS/Backend/Sources/Planbok/Features/FactorSource/Shields/New/Children/SetFactorThresholdFeature.swift
@@ -38,7 +38,7 @@ public struct SetFactorThresholdFeature {
 			guard numberOfFactors > 0 else {
 				return options
 			}
-			let exceeding1 = UInt16(numberOfFactors - 1)
+			let exceeding1 = UInt8(numberOfFactors - 1)
 			if exceeding1 > 1 {
 				options.append(contentsOf: (1...exceeding1).map(FactorThreshold.threshold))
 			}
